Remove support for redis-py < 4.2.
Remove aioredis support and tests
